+++ src/Settings_template.cpp
@@ -3,9 +3,9 @@
 
 
 ClockSettings clockConfigs[] = {
-    { "WifiClock",    0x00a2ff, 0x00a2ff, 0x00a2ff, 100, -50 } //0 light blue
+    { "WifiClock",    "Wifi name", "Wifi password", 0x00a2ff, 0x00a2ff, 0x00a2ff, 0xFF0000, 0xFF0000, 0xFF0000, 50, -150 }
 };
-int selectedClock = 0;  // Change this value to switch between clocks
+int selectedClock = 0;
 ClockSettings currentConfig;
 
 
@@ -14,17 +14,10 @@ const double LATITUDE = 52.3676; // Latitude for Amsterdam
 const double LONGITUDE = 4.9041; // Longitude for Amsterdam
 const double ZENITH = 90.833; // Standard Zenith for sunrise/sunset
 
-
-// Wi-Fi settings
-const char* ssid = "Wifi name";
-const char* password = "a nice password";
-
 //timeserver
 WiFiUDP ntpUDP;
 const char* ntpServerName = "pool.ntp.org";
 NTPClient timeClient(ntpUDP, ntpServerName, 0);
-
-
 
 
 /////////////////////////
@@ -37,14 +30,14 @@ uint32_t minuteColor = currentConfig.DEFAULT_MINUTE_COLOR;
 uint32_t colonColor = currentConfig.DEFAULT_COLON_COLOR;
 
 //brightness offset
-int dayTimeBrightnessOffset = currentConfig.brightnessOffsetDay;
-int nightTimeBrightnessOffset = currentConfig.brightnessOffsetDay;
+int dayTimeBrightnessOffset = currentConfig.dayTimeBrightnessOffset;
+int nightTimeBrightnessOffset = currentConfig.nightTimeBrightnessOffset;
 
 
 // Define the actual variables
 ClockMode currentMode = ClockMode::CLOCK_MODE;    // Initialize with a valid ClockMode value
 unsigned long modeStartTime = 0;
 bool isDaytime = true;
-const char* softwareVersion = "";
+const char* softwareVersion = "0.210";
 
 */